Add signal classifier

+++ flashing-jacket/src/classifiers/gnb_classifier.cpp
@@ -6,10 +6,10 @@
 GnbClassifier::GnbClassifier() {
 }
 
-int GnbClassifier::predict(const float *point) {
-	int cls = 0;
+uint8_t GnbClassifier::predict(const float *point) {
+	uint8_t cls = 0;
 	float dist = -FLT_MAX;
-	for (int i = 0; i < getNbClasses(); i++) {
+	for (uint8_t i = 0; i < getNbClasses(); i++) {
 		float new_distance = logLikelihood(i, point);
 		if (new_distance > dist) {
 			dist = new_distance;
@@ -20,9 +20,9 @@ int GnbClassifier::predict(const float *point) {
 	return cls;
 }
 
-float GnbClassifier::logLikelihood(const int cls, const float *point) {
+float GnbClassifier::logLikelihood(const uint8_t cls, const float *point) {
 	float dist = log(getClassPrior(cls));
-	for (int i = 0; i < getNbDims(); i++) {
+	for (uint8_t i = 0; i < getNbDims(); i++) {
 		dist -= 0.5 * ( log(2 * M_PI * getStdev(cls, i)) );
 		dist -= 0.5 * ( pow(getMean(cls, i) - point[i], 2) / getStdev(cls, i) );
 	}
