+++ flashing-jacket/src/classifiers/lda.cpp
@@ -5,9 +5,9 @@ LDA::LDA() {
 
 
 void LDA::transform(const float *point, float *transformed) {
-	for (int i = 0; i < getNbDims(); i++) {
+	for (uint8_t i = 0; i < getNbDims(); i++) {
 		transformed[i] = 0;
-		for (int j = 0; j < getNbFeatures(); j++) {
+		for (uint8_t j = 0; j < getNbFeatures(); j++) {
 			transformed[i] += (point[j] - getXbar(j)) * getScaling(j, i);
 		}
 	}
