+++ flashing-jacket/src/classifiers/lda.h
@@ -1,6 +1,8 @@
 #ifndef LDA_H
 #define LDA_H
 
+#include <stdint.h>
+
 class LDA {
 
 public:
@@ -10,11 +12,11 @@ public:
 
 protected:
 
-	virtual int getNbDims() = 0;
-	virtual int getNbFeatures() = 0;
+	virtual uint8_t getNbDims() = 0;
+	virtual uint8_t getNbFeatures() = 0;
 
-	virtual float getXbar(int feature) = 0;
-	virtual float getScaling(int feature, int cls) = 0;
+	virtual float getXbar(uint8_t feature) = 0;
+	virtual float getScaling(uint8_t feature, uint8_t cls) = 0;
 
 };
