+++ flashing-jacket/src/leds/signals/brake_signal.cpp
@@ -4,7 +4,7 @@
 
 static uint32_t BRAKE_COLOR = Adafruit_NeoPixel::Color(255,0,0);
 
-BrakeSignal::BrakeSignal(int nbRows, int nbColumns) : Signal(nbRows, nbColumns) {
+BrakeSignal::BrakeSignal(uint8_t nbRows, uint8_t nbColumns) : Signal(nbRows, nbColumns) {
 
 }
 
@@ -12,6 +12,6 @@ void BrakeSignal::refresh() {
 	
 }
 
-uint32_t BrakeSignal::getColor(int index) {
+uint32_t BrakeSignal::getColor(uint16_t index) {
 	return BRAKE_COLOR;
 }
