+++ flashing-jacket/src/leds/signals/brake_signal.h
@@ -5,10 +5,10 @@
 
 class BrakeSignal: public Signal {
 public:
-	BrakeSignal(int nbRow, int nbColumns);
+	BrakeSignal(uint8_t nbRow, uint8_t nbColumns);
 
 	virtual void refresh() override;
-	virtual uint32_t getColor(int index) override;
+	virtual uint32_t getColor(uint16_t index) override;
 
 };
