+++ flashing-jacket/src/leds/signals/direction_signal.cpp
@@ -5,7 +5,7 @@
 static uint32_t DIRECTION_COLOR = Adafruit_NeoPixel::Color(255,255,0);
 static uint32_t OFF_COLOR = Adafruit_NeoPixel::Color(0,0,0);
 
-DirectionSignal::DirectionSignal(int nbRows, int nbColumns, int direction): Signal(nbRows, nbColumns), mDirection(direction) {
+DirectionSignal::DirectionSignal(uint8_t nbRows, uint8_t nbColumns, int8_t direction): Signal(nbRows, nbColumns), mDirection(direction) {
 	mCounter = 0;
 }
 
@@ -13,9 +13,9 @@ void DirectionSignal::refresh() {
 	mCounter = mod(mCounter + mDirection, mNbRows);
 }
 
-uint32_t DirectionSignal::getColor(int index) {
-	int row = getRow(index);
-	int column = getColumn(index);
+uint32_t DirectionSignal::getColor(uint16_t index) {
+	uint8_t row = getRow(index);
+	uint8_t column = getColumn(index);
 
 	row = mod(row - mCounter, mNbRows);
