+++ flashing-jacket/src/leds/signals/idle_signal.cpp
@@ -4,7 +4,7 @@
 
 static uint32_t IDLE_COLOR = Adafruit_NeoPixel::Color(0,255,0);
 
-IdleSignal::IdleSignal(int nbRows, int nbColumns): Signal(nbRows, nbColumns) {
+IdleSignal::IdleSignal(uint8_t nbRows, uint8_t nbColumns): Signal(nbRows, nbColumns) {
 
 }
 
@@ -12,6 +12,6 @@ void IdleSignal::refresh() {
 }
 
 
-uint32_t IdleSignal::getColor(int index) {
+uint32_t IdleSignal::getColor(uint16_t index) {
 	return IDLE_COLOR;
 }
