+++ flashing-jacket/src/leds/signals/idle_signal.h
@@ -5,10 +5,10 @@
 
 class IdleSignal: public Signal {
 public:
-	IdleSignal(int nbRows, int nbColumns);
+	IdleSignal(uint8_t nbRows, uint8_t nbColumns);
 
 	virtual void refresh() override;
-	virtual uint32_t getColor(int index) override;
+	virtual uint32_t getColor(uint16_t index) override;
 };
 
 #endif
